Small Refactor
- Moved from team to contributors as the URL
- Added the route to header & footer, instead of only footer and updated the name from "Team" to "Contributors" as this is an OSS project and does not contain a dedicated resource on it
- Refactored contributors/page.tsx to be a bit more internally consistent with the rest of the application in regards to location
- Moved description to config/contributors.ts and will later expand. This is for i18n standardisation later on
