+++ frontend/config/contributors.ts
@@ -1,0 +1,5 @@
+export type ContributorsConfig = typeof contributorsConfig
+
+export const contributorsConfig = {
+    description: "The development of makeread.me could not have been possible without the help of our contributors. Here's a little thanks to the people that have helped along the way!",
+}
